next

+++ macsetup.sh
@@ -1,39 +1,49 @@
-# Take ownership
-sudo chown -R $USER /usr/local/
+# Let us not forget
+read -p "Did you do the steps in macsetup_readme?(y/n) " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
 
-# Get rid of unwanted Python
-#sudo rm -rf /Library/Frameworks/Python.framework
+  # Switch to z-shell
+  curl -L https://github.com/robbyrussell/oh-my-zsh/raw/master/tools/install.sh | sh
 
-# Check for Homebrew,
-# Install if we don't have it
-if test ! $(which brew); then
-  echo "Installing homebrew..."
-  ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
-fi
+  # Get rid of unwanted Python
+  #sudo rm -rf /Library/Frameworks/Python.framework
 
-# Update homebrew recipes
-brew update
+  # Check for Homebrew,
+  # Install if we don't have it
+  if test ! $(which brew); then
+    echo "Installing homebrew..."
+    ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+  fi
 
-# Install GNU core utilities (those that come with OS X are outdated)
-brew install coreutils
+  # Take ownership
+  sudo chown -R $USER /usr/local/
 
-# Install GNU `find`, `locate`, `updatedb`, and `xargs`, g-prefixed
-brew install findutils
+  # Update homebrew recipes
+  brew update
 
-# Install Bash 4
-brew install bash
+  # Install GNU core utilities (those that come with OS X are outdated)
+  brew install coreutils
 
-# Install more recent versions of some OS X tools
-brew tap homebrew/dupes
-brew install homebrew/dupes/grep
+  # Install GNU `find`, `locate`, `updatedb`, and `xargs`, g-prefixed
+  brew install findutils
 
-#$PATH=$(brew --prefix coreutils)/libexec/gnubin:$PATH
+  # Install Bash 4
+  brew install bash
 
-# Vim
-brew install macvim --with-override-system-vim
-brew linkapps
+  # Install more recent versions of some OS X tools
+  brew tap homebrew/dupes
+  brew install homebrew/dupes/grep
 
-binaries=(
+  #$PATH=$(brew --prefix coreutils)/libexec/gnubin:$PATH
+
+  # Vim
+  brew install macvim --with-override-system-vim
+  brew linkapps
+
+  binaries=(
+  cabextract
   cmake
   #graphicsmagick
   #webkit2png
@@ -49,17 +59,17 @@ binaries=(
   #hub
   git
   mackup
-)
+  )
 
-echo "installing binaries..."
-brew install ${binaries[@]}
+  echo "installing binaries..."
+  brew install ${binaries[@]}
 
-brew cleanup
+  brew cleanup
 
-# Apps
-brew install caskroom/cask/brew-cask
+  # Apps
+  brew install caskroom/cask/brew-cask
 
-apps=(
+  apps=(
   adobe-creative-cloud
   alfred
   dropbox
@@ -97,39 +107,46 @@ apps=(
   #quicklook-json
   skype
   #transmission
-)
+  )
 
-# Install apps to /Applications
-# Default is: /Users/$user/Applications
-echo "installing apps..."
-brew cask install --appdir="/Applications" ${apps[@]}
+  # Install apps to /Applications
+  # Default is: /Users/$user/Applications
+  echo "installing apps..."
+  brew cask install --appdir="/Applications" ${apps[@]}
 
-# Link Alfred
-brew cask alfred link
+  # Link Alfred
+  brew cask alfred link
 
-# Fonts
-brew tap caskroom/fonts
+  # Fonts
+  brew tap caskroom/fonts
 
-fonts=(
+  fonts=(
   font-m-plus
   font-clear-sans
   font-roboto
-)
+  font-source-code-pro
+  font-inconsolata
+  )
 
-# Install fonts
-echo "Installing fonts..."
-brew cask install ${fonts[@]}
+  # Install fonts
+  echo "Installing fonts..."
+  brew cask install ${fonts[@]}
 
-# Setup
-sh "scripts/osx-for-hackers.sh"
+  # Setup
+  sh "scripts/osx-for-hackers.sh"
 
-echo "Installing YouCompleteMe"
-cd ~
-mkdir ycm_build
-cd ycm_build
-cmake -G "Unix Makefiles" . ~/.vim/bundle/YouCompleteMe/third_party/ycmd/cpp
-cmake --build . --target ycm_support_libs --config Release
-echo "YouCompleteMe install successful."
+  cd ~/Downloads
+  mkdir consolas
+  cd consolas
+  curl -O http://download.microsoft.com/download/f/5/a/f5a3df76-d856-4a61-a6bd-722f52a5be26/PowerPointViewer.exe
+  cabextract PowerPointViewer.exe
+  cabextract ppviewer.cab
+  open CONSOLA*.TTF
 
-echo "Done. Remember to login to Dropbox and run 'mackup restore' for settings. Disabling Caps Lock in Keyboard settings may be necessary."
+  echo "Click Install font."
+  echo "Then login to Dropbox and run"
+  echo "mackup restore"
+  echo "Once restored, run"
+  echo "sh YouCompleteMe.sh"
 
+fi
